pyproject.toml
[tool.poetry]
name = "nftdummies-bot"
version = "0.1.0"
description = "A bot for @nftdummies"
authors = ["welaskez <nazar.nikitenko.03@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.8.0"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
pytonapi = "^0.3.2"
aiosqlite = "^0.20.0"
sqlalchemy = "^2.0.31"
greenlet = "^3.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
